# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "This is Wasilny Api Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/user/auth/login": {
            "post": {
                "tags": [
                    "User Auth"
                ],
                "summary": "الدخول للنظام",
                "description": "",
                "operationId": "Login",
                "requestBody": {
                    "description": "جميع الخيارات إلزامية",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "String",
                                        "format": "966536301031",
                                        "example": "966536301031"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "FALSE"
                                        },
                                        "status": {
                                            "type": "int",
                                            "example": 200
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": {
                                                "message": "OK"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/auth/verfyOTP": {
            "post": {
                "tags": [
                    "User Auth"
                ],
                "summary": "تحقق من الرقم",
                "description": "",
                "operationId": "OTP",
                "requestBody": {
                    "description": "جميع الخيارات إلزامية",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "String",
                                        "format": "966536301031",
                                        "example": "966536301031"
                                    },
                                    "otp": {
                                        "type": "String",
                                        "format": "1111",
                                        "example": "1111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "502": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "FALSE"
                                        },
                                        "status": {
                                            "type": "int",
                                            "example": 502
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": {
                                                "message": "error message"
                                            }
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/uploadImage": {
            "post": {
                "tags": [
                    "Driver Auth"
                ],
                "summary": "رفع صورة مستندات للسائق",
                "description": "",
                "operationId": "imageUploadPost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "file to upload",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "FALSE"
                                        },
                                        "status": {
                                            "type": "int",
                                            "example": 502
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": {
                                                "imageName": "image.png"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}